Deterministic Codeflow
Modified code to achieve deterministic codeflow, but it took a lot of
debugging and performance is now slow. Seems the problem is with sets
being arranged differently in memory run to run, so even if we randomly
select element X every run, the element in position X isn't necessarily
the same. Code needs some cleanup and optimization after this checkin.